+++ api/src/routes/index.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const staticLayers = require('./staticLayers');
+const dataLayers = require('./dataLayers');
 const attributes = require('./attributes');
 const pointAttributes = require('./pointAttributes');
 const uploads = require('./uploads');
@@ -8,7 +8,8 @@ const config = require('./config');
 
 const router = express.Router();
 
-router.use('/staticLayers', staticLayers);
+router.use('/dataLayers', dataLayers);
+router.use('/staticLayers', dataLayers);
 router.use('/attributes', attributes);
 router.use('/pointAttributes', pointAttributes);
 router.use('/uploads', uploads);