+++ api/src/tests/attributes.js
@@ -1,6 +1,6 @@
 const request = require('supertest');
 const app = require('../config/express');
-const { Attribute, DataDateAttribute, NumberAttribute } = require('../dbSchemas/attributeSchema');
+const { Attribute, DataDateAttribute, NumberAttribute } = require('../database/mongoDb/dbSchemas/attributeSchema');
 const {
   casesAttrToDb,
   deathsAttrToDb,
@@ -13,6 +13,7 @@ const {
 jest.mock('../config/config.js', () => {
   return {
     authorizeTokenAttribute: false,
+    mongoUri: 'qwertyuiop',
   };
 });