+++ initial-data-load/src/mapLayers/layersGeoDataUpload.js
@@ -4,21 +4,20 @@ const yaml = require('js-yaml');
 const axios = require('axios');
 const bbox = require('@turf/bbox');
 
+const config = require('../config/config');
 const logger = require('../config/winston');
 
-const {
-  getOneLayerGeoData,
-  saveLayerGeoData,
-  storeGeoFeaturesData,
-  createGeoDataIndex,
-  deleteAllFromCollection,
-} = require('./db');
-const { saveGeoJsonFromUrlSourceToStorage, saveGeoJsonFromFileToStorage } = require('./storage');
+const { createCollection } = require('../database');
+const { getOneLayerGeoData, saveGeoData } = require('../database/layers');
+const { storeGeoFeaturesData } = require('../database/geoFeatureCollections');
+
+const { storeLocalFileAsBlob, storeFromUrlAsBlob } = require('../azureStorage/blobContainer');
+const { clearCollection } = require('../database');
 
 const storeGeoDataToDb = async (fromFile, data, filePath) => {
   let geojsonData;
   logger.info(`Clearing database collection ${data.collectionName}`);
-  await deleteAllFromCollection(data.collectionName);
+  await clearCollection(data.collectionName);
   logger.info(`Loading data from file ${filePath}`);
   if (fromFile) {
     geojsonData = JSON.parse(fs.readFileSync(filePath, 'utf8'));
@@ -31,7 +30,7 @@ const storeGeoDataToDb = async (fromFile, data, filePath) => {
     geojsonData = result.data;
   }
   logger.info(`Creating index on collection...`);
-  await createGeoDataIndex(data.collectionName, 'bbox');
+  await createCollection(data.collectionName, undefined, 'bbox');
   logger.info(`Generating bounding boxes for features...`);
 
   const { features } = geojsonData;
@@ -64,7 +63,7 @@ const storeGeoDataToDb = async (fromFile, data, filePath) => {
   });
 
   logger.info(`Storing features into database...`);
-  await storeGeoFeaturesData(featuresWithBbox, data.collectionName);
+  await storeGeoFeaturesData(data.collectionName, featuresWithBbox);
 };
 
 const formatLayerGeoData = async (data, country) => {
@@ -75,7 +74,7 @@ const formatLayerGeoData = async (data, country) => {
   }
   let url;
   if (data.collectionName && data.apiUrl) {
-    await createGeoDataIndex(data.collectionName);
+    await createCollection(data.collectionName, undefined, 'geometry');
     url = data.apiUrl;
   }
   if (data.geoDataUrl) {
@@ -85,7 +84,7 @@ const formatLayerGeoData = async (data, country) => {
       url = data.apiUrl;
       logger.info(`Layer ${data.name} has new URL for geodata: ${url}`);
     } else {
-      const filename = await saveGeoJsonFromUrlSourceToStorage(data.geoDataUrl);
+      const filename = await storeFromUrlAsBlob(data.geoDataUrl, config.azureStorageLayerContainerName);
       if (filename) {
         url = `/api/uploads/geojsons/${filename}`;
         logger.info(`Layer ${data.name} has new URL for geojson: ${url}`);
@@ -101,7 +100,11 @@ const formatLayerGeoData = async (data, country) => {
       logger.info(`Layer ${data.name} has new URL for geodata: ${url}`);
     } else if (hasFile && !data.storeToDb) {
       logger.info(`Storing geojson ${data.geoDataFilename} for layer ${data.referenceId} from file...`);
-      const filename = await saveGeoJsonFromFileToStorage(filePath, data.geoDataFilename);
+      const filename = await storeLocalFileAsBlob(
+        filePath,
+        data.geoDataFilename,
+        config.azureStorageLayerContainerName,
+      );
       if (filename) {
         url = `/api/uploads/geojsons/${filename}`;
         logger.info(`Layer ${data.name} has new URL for geojson: ${url}`);
@@ -129,7 +132,7 @@ const geoDataUpload = async (country) => {
       const savingGeoJsons = layersGeoData.map((item) => formatLayerGeoData(item, country));
       let geoDataForDb = await Promise.all([...savingGeoJsons]);
       geoDataForDb = geoDataForDb.filter((item) => item);
-      await saveLayerGeoData(geoDataForDb);
+      await saveGeoData(geoDataForDb);
       logger.info(`LayerGeoData for country ${country} successfully saved to database.`);
     } else {
       logger.error(`Data for country ${country} not found.`);