+++ initial-data-load/src/testUtils/setup.js
@@ -1,4 +1,4 @@
-const { initializeDBConnection, disconnectFromDB, removeDB } = require('../db');
+const { initializeDBConnection, disconnectFromDB, removeDB } = require('../database/mongoDb/index');
 
 // the default jest timeout is 5s which is sometimes not enough to initialize DB connection
 const TIMEOUT = 30000;