+++ initial-data-load/src/tests/attributes.js
@@ -6,6 +6,7 @@ describe('Layer attributes', () => {
   it('should save nothing, if data for country not found', async () => {
     jest.mock('../config/config.js', () => {
       return {
+        mongoUri: 'qwertyuiop',
         uploadDataTypes: 'attributes',
         country: 'testCountry2',
       };
@@ -25,6 +26,7 @@ describe('Layer attributes', () => {
   it('should upload data with correct date', async () => {
     jest.mock('../config/config.js', () => {
       return {
+        mongoUri: 'qwertyuiop',
         uploadDataTypes: 'attributes',
         country: 'testCountry',
       };